Fixed a comment
Sander has stolen my sanity, I don't think I'll be the same man after this.
We reverted the revert, and for what? Just to create another pull-request?
Is this what my life has resorted too? Pull-requests for a single comment change?
Was it worth it? Was all this pain and suffering worth it? Just for a single correction in a comment?
I hope it is, as it has costed me my life. And with that, my sanity.
I am a changed man. I can never be the same.
Goodbye.